Dios dame fuerza

+++ src/main/java/com/uwu/cliente/Entity/producto.java
@@ -15,13 +15,13 @@ public class producto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_producto")
-    private long id;
+    private Long id;
     private String nombre;
     private String descripcion;
     @PositiveOrZero(message = "PUES DEJEMOSLO GRATIS PERO NI MODO QUE LES DEMOS PLATA")
-    private double p_unitario;
+    private float p_unitario;
     @PositiveOrZero(message = "COMO VAMOS A VENDER ALGO QUE NO TENEMOS Y ADEMAS DEBEMOS (SI SOY)")
-    private int stock;
+    private float stock;
     private boolean estado = true;
 
     //@OneToMany(mappedBy = "producto")
@@ -30,7 +30,7 @@ public class producto {
     public producto() {
     }
 
-    public producto(long id, String nombre, String descripcion, double p_unitario, int stock, boolean estado) {
+    public producto(Long id, String nombre, String descripcion, float p_unitario, float stock, boolean estado) {
         this.id = id;
         this.nombre = nombre;
         this.descripcion = descripcion;
@@ -40,11 +40,11 @@ public class producto {
         //this.detalles = detalles;
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -64,19 +64,19 @@ public class producto {
         this.descripcion = descripcion;
     }
 
-    public double getP_unitario() {
+    public float getP_unitario() {
         return p_unitario;
     }
 
-    public void setP_unitario(double p_unitario) {
+    public void setP_unitario(float p_unitario) {
         this.p_unitario = p_unitario;
     }
 
-    public int getStock() {
+    public float getStock() {
         return stock;
     }
 
-    public void setStock(int stock) {
+    public void setStock(float stock) {
         this.stock = stock;
     }
